platform/data-pipeline/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
minio==7.2.0
prometheus-client==0.19.0
asyncio==3.4.3
python-dateutil==2.8.2